backend/requirements.txt
python-dotenv==1.0.0
rembg==2.0.50
onnxruntime==1.16.3
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
yt-dlp==2024.1.7
requests==2.31.0
gunicorn==21.2.0

